Close and Reopen log files across a checkpoint/restore
A common best practice in CRIU mode, at least in OpenJ9, is to close all
open files before a checkpoint, and reopen them on restore. The JIT
would keep log files such as the vlog and rtLog open across a
checkpoint/restore boundary. This commit closes these files on
checkpoint and reopens them on restore.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>