Retry failed profiled very-hot compilations at hot instead of warm
Before this change, a profiled-very-hot compilation that fails
due to excessiveComplexity would not be retried, relying on the
existing compiled body. If the existing compiled body is at
warm (or even cold) opt level, then the performance could be
suboptimal.

With this commit profiled-very-hot compilations that fail will be
retried at the hot optimization level, but without any profiling.
We avoid profiling compilations because they are more complex than
non-profiled ones and the likelihood of a failure for a
profiled-hot compilation is relatively high when we know that a
profiled-very-hot compilation for the same method already failed.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>