Fix x86-32 private linkage for dispatchVirtual
Previously, the VTable index argument register was not marked as such on
x86-32 for 32-bit arguments, which caused issues when invoking
MethodHandles through dispatchVirtual.

Now, the Vtable index register is marked (by a precondition) for
integer arguments. Additionally, the special handling required for
dispatchVirtual is moved outside of the argument setup loop to improve
clarity.

Signed-off-by: Matthew Hall <matthew.hall3@outlook.com>